refactor: Move API method from content fields to hook.
