+++ src/admin/pages/collections/Context/types.ts
@@ -1,6 +1,6 @@
 import { SWRConfiguration, SWRResponse } from 'swr';
 import { SWRMutationResponse } from 'swr/mutation';
-import { Field } from '../../../../config/types.js';
+import { Field, File } from '../../../../config/types.js';
 
 export type ContentContext = {
   getContents: (canFetch: boolean, slug: string, config?: SWRConfiguration) => SWRResponse<any[]>;
@@ -17,4 +17,11 @@ export type ContentContext = {
     slug: string,
     id: string
   ) => SWRMutationResponse<void, any, Record<string, any>, any>;
+  getFileImage: (id: string | null) => SWRMutationResponse<{ file: File; raw: string }>;
+  createFileImage: () => SWRMutationResponse<
+    { file: File; raw: string },
+    any,
+    Record<string, any>,
+    any
+  >;
 };
